split out functionality into separate files
